V1.1 Added WeekMenu to GTVultIn applicatie
